requirements.txt
Django==3.2.17
celery==5.2.7
djangorestframework-jwt==1.11.0
djangorestframework==3.14.0
django_celery_beat==2.4.0
django_celery_results==2.4.0
django-cors-headers==3.13.0
django-filter==22.1
django-redis==5.2.0
django-silk==5.0.3
drf-yasg==1.21.4
gunicorn==20.1.0
mongoengine==0.27.0
mysqlclient==2.1.1
psycopg2-binary==2.9.5
python-decouple==3.7
python-json-logger==2.0.7
redis==4.4.2
sentry-sdk==1.19.1
whitenoise~=6.3.0
